PR Feeback: Move things to individual namespaced files.
