+++ guardrails/call_tracing/__init__.py
@@ -1,3 +1,7 @@
 """
 For tracing (logging) and reporting the timing of Guard and Validator calls.
 """
+from guardrails.call_tracing.trace_entry import GuardTraceEntry
+from guardrails.call_tracing.guard_call_logging import TraceHandler
+
+__all__ = ['GuardTraceEntry', 'TraceHandler']
